Started Placeability
We have lowered the scope of the project but we still need to work on
all the placeable components. A placeableComponents file was created so
we can work on them. We should finish the documentation now to use it as
a guide. Also test.js should stay at is, and we should code around it.